requirements.txt
apache-airflow==2.10.3
apache-airflow-providers-openlineage>=1.8.0,<2.0.0
apache-airflow-providers-apache-spark>=4.1.0,<5.0.0
delta-spark==3.0.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.1
pyarrow==14.0.1
apache-airflow-providers-openlineage>=1.8.0
awscli
pyspark

